TRDRYDEP.f:      Removed Olson PFT formatted input file dependence, instead using the land fraction and leaf area index data from Ent, including the necessary mapping from Ent to DryDep vegetation types (encoded in IDEP array). The previous three land fraction arrays (IJLAND, IJUSE, IJREG) are now simplified to only IJUSE which encodes the land type and fraction data. Removed factor of 1000 from IJUSE. Hardcoded the polynomial cofficients, maximum deposition velocity, resistances, and Ent->DryDep PFT mapping into RDDRYCF. Routines READLAI and RDLAND are now obsolete and were thus removed.
BIOGENIC_EMISSIONS.f:      Base isoprene emission factors in BIOGENIC_EMISSIONS were not known for the Ent types, so a mapping from Ent to Megan types (for which emission factors are known) is performed using Greg Faluvegi's map_ent_pfts_to_megan_pfts routine. I moved this mapping routine from megan to ENT_DRV so that it can be used even when megan is not compiled. This routine now outputs an array which encodes the mapping  so that it can be used in BIOGENIC_EMISSIONS. Unit conversion from one set of emissions factors to the other is now included in the "factor" variable in BIOGENIC_EMISSIONS. No tuning of the base isoprene emissions was included, and the isoprene emissions are generally increased, with the geographic pattern being roughly similar to before.

GHY_DRV.f:      Moved calls to RDISOBASE, RDISOPCF, RDLAI, and RDDRYCF from initTracerGriddedData to GHY_DRV so that the Ent vegetation data can be utilized. Removed calls to obsolete routines.

megan.F90:      map_ent_pfts_to_megan_pfts routine was moved to ENT_DRV and nMeganPFT parameter was moved to TRCHEM_Shindell_COM so that they can be used even when megan is not compiled.

Templates:      Removed inclusion of dry_depos_144x90_input_files as these input files are no longer used.