Wind nudging code can now determine the nudging time interval automatically from the wind file time dimensions, allowing for the use of three-hourly and shorter time interval wind fields without needing to modify the run-deck file.  Also re-factored the nudging code slightly to improve model throughput.
Finally, commented-out instructions on using the MERRA2 six-hourly and
three-hourly nudging winds were added to the nudging rundeck template. If one
is interested in the three-hourly MERRA2 nudging winds, then one can find them
on discover here:

/discover/nobackup/projects/giss/OBS/giss_nudging_winds/MERRA2/three_hourly

Please note that if using the default nudging relaxation time-scale of ~16
minutes, one may see little to no difference when using the higher-frequency
winds.  However, it could certainly have an influence if using time-scales on
the order of hours, which seems much more commonplace in the literature.