This comits adds functionality to read time-varying vegetation cover data from VEG file.
Old functioality didn't change and old rundecks and VEG files work
as before.

To provide time-varying data, VEG file should have a format similar
to CROPS file. The difference is that instead of 1 variable it should
contain 16 variables corresponding to cover types. Missing variables
are assumed to be zero. The behavior is controlled by the same
crops_yr rundeck parameter:

 crops_yr = 0 ! use time-varying data corresponding to model year
 crops_yr = XXXX ! use data for the year XXXX
 crops_yr = -1 ! is an error: crops have to be provided by CROPS file

As it is mentions above, this functionality requires that crops
are provided by CROPS file. If CROPS file is missing, crops will
be set to zero.