minor bug fixes + WIP MEGAN updates
routine map_ent_pfts_to_megan_pfts was expanded to have two optional
(but positional) arguments:

(1) x_ent_in - That is, if another variable of dimension {num ent
    pft types} is passed after the i, and j, then instead of
    returning the pft fraction per megan pft types, it returns that
    variable over megan pft types (in a weighted average sense)
    Since adding this functionality increased the repeated code in
    this routine, a convenience subroutine was added ("fill_it").

(2) meg - this argument already existed but it is now optional and
    x_ent_in above pushes it one slot to the right. Tyler C was
    using it to return the Ent-->megan mapping itself so that it
    could be utilized in the calling routine. (If more optional
    arguments are added, we should probably start using
    named/keyword ones instead.)

As a consequence of (1) and (2) the existing calls have been changed:
In BIOGENIC_EMISSIONS.f, where it had be used just to get the mapping,
an additional dummy array was needed. BUGFIX: Also changing megan_map
to be integers here, as Andy A. already corrected in the cloud sandbox.
Conversely, in megan.F90, the dummy array in the call is no longer
needed.

megan.F90: BUGFIX: The gamma for CO2 used to be set outside of the loop
over species, but reset inside the loop if current species is not
isoprene-related. It could therefore be wrong if there were an
isoprene-related species later in the loop. Turns out, as coded, all
the isoprene was done first, so this should have had no effect. But now
it's fixed by setting gamma_CO2 inside the loop always.

megan.F90 development: Currently the parts of code that would change the
answer (Elizabeth's work) are protected inside #defined KLOVENSKI_DEV
directives. So far these are preparations to use maximum photosynthetic
capacity (vcmax-related stuff; either from Ent or a look-up table) and
water stress (from Ent) to provide a non-unity soil moisture gamma. The
soil moisture gamma calculation was therefore moved into the species
loop, as this is only non-unity for isoprene. In fact it uses a new
routine 'get_gamma_sm' the previous 'get_gamma_s' is left in as comments
for now.

Unrelated BUGFIX: TRCHEM_fastj2.f altered an if statement to prevent
trap problems when no aerosol aods are defined (Andy A. fix on cloud
sandbox; this wasn't hurting anything, as we were never running
chemistry with no aerosol scheme.)