add VARIABLE_COLMO3_FROM_NINT option
(No change in results, as this new CPP directive is not on
in any templates.) Story: Shindell chemistry has an assumed
O3 amount above the top of the chemistry that goes into the
"SF2 coding" which affects NO-->N+O photolysis. This is the
initial value of the "colmo3" variable. It is a simple hard-
code (though these days scaled by the model top pressure).
On the E2.2 merge branch we have another scaling by the
local (I,J) O3 at top of chemistry compared to a global mean
at that level. But putting in global means in the code makes
people angry. So on this branch I am trying out using the
NINT (above chemistry layers) O3 input amount in lieu of the
hardcode and non-local scaling.

This can be considered experimental, which is why it's in a
CPP. Turns out the amount of ozone in the initial colmo3 is
quite different, if you use this, but it deosn't seem to
affect the top-of-chemistry O3 or Ox much. Someday we'll test
on E3 with 102L.