The parameter "keep_params" was added to allow an ISTART=8 start that gives the same results as a continuation of the run (ISTART=9). To achieve that, you have to add in the PARAMS section:
keep_params=1
init_topog_related=0
init_flake=0
niprnt=1

The last line is only needed if you want to print out the initial conditions
and create the awkward PARTIAL.acc file.

Another condition for identical results is that the radiation is called at the
same times as in the original run; this is automatic if nrad=1, with nrad=5 it
is automatic for a 365-day year if we start at the same calendar day as the
original run.

Note that only the prognostic variables have the same values, some diagnostics
may be slightly different (unless all diagnostics were called at that starting
point also in the original run).