Added logic to permit opt-in instantiation of each member of the TRACERS_PASSIVE collection, via the rundeck.   For the time being, the list of tracers is passed via the _value_ of the TRACERS_PASSIVE CPP variable.    Example:
\#define TRACERS_PASSIVE "aoa st8025"

(leading backslash in this commit message so git does not interpret # as a comment)

If no value is specified, the entire collection is taken, reproducing the previous behavior.

CPP was used instead of a rundeck "parameter" to signal that this logic is only intended
as a compile-time flexibility right now.  Full runtime flexibility can only happen as
part of a more systematic solution involving all the other atmospheric tracer collections
as well, and will be more work than the instantiation step (restarts and other scenarios
must be handled).