CH4 emissions tuning prep for SSPs
Since this tuning has been more challanging than CMIP5 version was,
(partially b/c we are targeting a longer obsercational timeseries
and partially because prescribed emissions vary less smoothly now),
we may continue to do tests until the landuse is updated in this
template. But thought I would push the settings to use if runs
were to start today.

Normally we'd tune wetlands, but since the model seemed to want to
tune this upward, and since it seemed to us our wetlands source
is on the high side of estimates, we instead ended up tuning a
smaller source that seems highly uncertain and where we may be low:
the ocean/lake/misc ground source.

So this push does: wetlands*1.0+misc*2.0
(technically, wetlands also tuned upward as CMIP5 had 0.9)