This is a hack which allows a call stack dumping upon the model crash with the compiles which refuse to do it by themselves. It dumps the call stack only for the floating point exception and it relies on gdb debugger being present on the system (gdb-ia for Intel).
This hack is not very portable. It is expected to work on discover
cluster with Intel 20 (19.1) compiler. It *may* work on a typical
Linux desktop with gfortran compiler.

To enable it, define USE_GDB_FOR_FPE_BACKTRACE in your rundeck.
Keep in mind, that the call stack will appear at the end of your
*.PRT file and not in any of the system logs.