SLES12 update OMA-->NINT input script
First, a note on reproducibility or lack there-of of the
NINT input files created:

I used this utility / version:
mkelley5 1477606 Nov 20  2014 /discover/nobackup/projects/giss/exec/diffreport
on 2020.18.03 to test a 10 year batch of OMA-->NINT input files
created with the new script(s) committed here vs. already-created
input files for CMIP6 Larissa and I made on SLES11.
I compared both the year-to-year input files and the 10-year average
"climatology" input. The yearly input files were identical. I believe
these are what we've normally been using for CMIP6 NINT runs... The
climatology files had max relative differences on the order of 1E-7.
I believe this is because of a change in version for the NCO ncea (ensemble
averager) utility used to average the acc files. I.e. the version we used
on SLES11 is not available on SLES12. In other words, differences in
results do not appear to be due to what's being changed in this commit.

This commit:

The modules I tested on SLES12 are now explicitly loaded in the script,
and locations of needed utilities (like NCO) specified. This means they
are no longer passed from the calling script and thus the script has one
less argument. Also when utilities like ncdump or scaleacc are used, a
a specific full-path command, rather than whatever the user's environment
points to is utilized.

This is a ksh script, and the version on SLES12 (aliased to mksh) doesn't
support floating point comparisons. So in one place this task is offloaded
to linux 'bc' command.

In case blocks, removed semicolons with no preceding command. SLES11 ksh
was more permissive of that syntax.

Newer ncks version needed a -C flag to avoid certain "HINT" messages
to standard output. Few cosmetic cleanups.

The example calling script is also updated to no longer define/pass nco
location. It also now does a simpler example, mostly just because I had
the acc files 'at hand' to test it.